Create shared `DisputeDueByDate` component
