+++ client/payment-details/dispute-details/dispute-summary-row.tsx
@@ -6,9 +6,8 @@
 import React from 'react';
 import moment from 'moment';
 import HelpOutlineIcon from 'gridicons/dist/help-outline';
-import { __, _n, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { dateI18n } from '@wordpress/date';
-import classNames from 'classnames';
 
 /**
  * Internal dependencies
@@ -21,20 +20,13 @@ import { formatStringValue } from 'wcpay/utils';
 import { ClickTooltip } from 'wcpay/components/tooltip';
 import Paragraphs from 'wcpay/components/paragraphs';
 import { getDisputeDeductedBalanceTransaction } from 'wcpay/disputes/utils';
+import DisputeDueByDate from './dispute-due-by-date';
 
 interface Props {
 	dispute: Dispute;
-	daysRemaining: number;
 }
 
-const DisputeSummaryRow: React.FC< Props > = ( { dispute, daysRemaining } ) => {
-	const respondByDate = dispute.evidence_details?.due_by
-		? dateI18n(
-				'M j, Y, g:ia',
-				moment( dispute.evidence_details?.due_by * 1000 ).toISOString()
-		  )
-		: '–';
-
+const DisputeSummaryRow: React.FC< Props > = ( { dispute } ) => {
 	const disputeReason = formatStringValue(
 		reasons[ dispute.reason ]?.display || dispute.reason
 	);
@@ -105,34 +97,7 @@ const DisputeSummaryRow: React.FC< Props > = ( { dispute, daysRemaining } ) => {
 		{
 			title: __( 'Respond By', 'woocommerce-payments' ),
 			content: (
-				<span className="dispute-summary-row__response-date">
-					{ respondByDate }
-					<span
-						className={ classNames( {
-							'dispute-summary-row__response-date--urgent':
-								daysRemaining < 3,
-							'dispute-summary-row__response-date--warning':
-								daysRemaining < 7 && daysRemaining > 2,
-						} ) }
-					>
-						{ daysRemaining > 0 &&
-							sprintf(
-								// Translators: %s is the number of days left to respond to the dispute.
-								_n(
-									'(%s day left to respond)',
-									'(%s days left to respond)',
-									daysRemaining,
-									'woocommerce-payments'
-								),
-								daysRemaining
-							) }
-
-						{ daysRemaining === 0 &&
-							__( '(Last day today)', 'woocommerce-payments' ) }
-						{ daysRemaining < 0 &&
-							__( '(Past due)', 'woocommerce-payments' ) }
-					</span>
-				</span>
+				<DisputeDueByDate dueBy={ dispute.evidence_details.due_by } />
 			),
 		},
 	];
